Updated project page
The index page for the github page is pretty much finished.
It's not styled very nice, but it looks ok.
I'm going to add an endless spin of on of the hollow circle at the top right later.
Also a page of the available styles and their characters needs to exist.